Move things around in `substrate-wasm-builder` to clean up `lib.rs` for future changes, no business logic changes here
