+++ utils/wasm-builder/src/lib.rs
@@ -105,192 +105,18 @@
 //! install nightly-2020-02-20`, the wasm target needs to be installed as well `rustup target add
 //! wasm32-unknown-unknown --toolchain nightly-2020-02-20`.
 
-use std::{
-	env, fs,
-	io::BufRead,
-	path::{Path, PathBuf},
-	process::Command,
-};
+use std::{fs, path::Path};
 
 mod builder;
-mod prerequisites;
-mod wasm_project;
+mod env_vars;
+mod utils;
 
 pub use builder::{WasmBuilder, WasmBuilderSelectProject};
 
-/// Environment variable that tells us to skip building the wasm binary.
-const SKIP_BUILD_ENV: &str = "SKIP_WASM_BUILD";
-
-/// Environment variable that tells us whether we should avoid network requests
-const OFFLINE: &str = "CARGO_NET_OFFLINE";
-
-/// Environment variable to force a certain build type when building the wasm binary.
-/// Expects "debug", "release" or "production" as value.
-///
-/// When unset the WASM binary uses the same build type as the main cargo build with
-/// the exception of a debug build: In this case the wasm build defaults to `release` in
-/// order to avoid a slowdown when not explicitly requested.
-const WASM_BUILD_TYPE_ENV: &str = "WASM_BUILD_TYPE";
-
-/// Environment variable to extend the `RUSTFLAGS` variable given to the wasm build.
-const WASM_BUILD_RUSTFLAGS_ENV: &str = "WASM_BUILD_RUSTFLAGS";
-
-/// Environment variable to set the target directory to copy the final wasm binary.
-///
-/// The directory needs to be an absolute path.
-const WASM_TARGET_DIRECTORY: &str = "WASM_TARGET_DIRECTORY";
-
-/// Environment variable to disable color output of the wasm build.
-const WASM_BUILD_NO_COLOR: &str = "WASM_BUILD_NO_COLOR";
-
-/// Environment variable to set the toolchain used to compile the wasm binary.
-const WASM_BUILD_TOOLCHAIN: &str = "WASM_BUILD_TOOLCHAIN";
-
-/// Environment variable that makes sure the WASM build is triggered.
-const FORCE_WASM_BUILD_ENV: &str = "FORCE_WASM_BUILD";
-
-/// Environment variable that hints the workspace we are building.
-const WASM_BUILD_WORKSPACE_HINT: &str = "WASM_BUILD_WORKSPACE_HINT";
-
 /// Write to the given `file` if the `content` is different.
-fn write_file_if_changed(file: impl AsRef<Path>, content: impl AsRef<str>) {
+pub(crate) fn write_file_if_changed(file: impl AsRef<Path>, content: impl AsRef<str>) {
 	if fs::read_to_string(file.as_ref()).ok().as_deref() != Some(content.as_ref()) {
 		fs::write(file.as_ref(), content.as_ref())
 			.unwrap_or_else(|_| panic!("Writing `{}` can not fail!", file.as_ref().display()));
 	}
-}
-
-/// Copy `src` to `dst` if the `dst` does not exist or is different.
-fn copy_file_if_changed(src: PathBuf, dst: PathBuf) {
-	let src_file = fs::read_to_string(&src).ok();
-	let dst_file = fs::read_to_string(&dst).ok();
-
-	if src_file != dst_file {
-		fs::copy(&src, &dst).unwrap_or_else(|_| {
-			panic!("Copying `{}` to `{}` can not fail; qed", src.display(), dst.display())
-		});
-	}
-}
-
-/// Get a cargo command that compiles with nightly
-fn get_nightly_cargo() -> CargoCommand {
-	let env_cargo =
-		CargoCommand::new(&env::var("CARGO").expect("`CARGO` env variable is always set by cargo"));
-	let default_cargo = CargoCommand::new("cargo");
-	let rustup_run_nightly = CargoCommand::new_with_args("rustup", &["run", "nightly", "cargo"]);
-	let wasm_toolchain = env::var(WASM_BUILD_TOOLCHAIN).ok();
-
-	// First check if the user requested a specific toolchain
-	if let Some(cmd) = wasm_toolchain.and_then(|t| get_rustup_nightly(Some(t))) {
-		cmd
-	} else if env_cargo.is_nightly() {
-		env_cargo
-	} else if default_cargo.is_nightly() {
-		default_cargo
-	} else if rustup_run_nightly.is_nightly() {
-		rustup_run_nightly
-	} else {
-		// If no command before provided us with a nightly compiler, we try to search one
-		// with rustup. If that fails as well, we return the default cargo and let the prequisities
-		// check fail.
-		get_rustup_nightly(None).unwrap_or(default_cargo)
-	}
-}
-
-/// Get a nightly from rustup. If `selected` is `Some(_)`, a `CargoCommand` using the given
-/// nightly is returned.
-fn get_rustup_nightly(selected: Option<String>) -> Option<CargoCommand> {
-	let host = format!("-{}", env::var("HOST").expect("`HOST` is always set by cargo"));
-
-	let version = match selected {
-		Some(selected) => selected,
-		None => {
-			let output = Command::new("rustup").args(&["toolchain", "list"]).output().ok()?.stdout;
-			let lines = output.as_slice().lines();
-
-			let mut latest_nightly = None;
-			for line in lines.filter_map(|l| l.ok()) {
-				if line.starts_with("nightly-") && line.ends_with(&host) {
-					// Rustup prints them sorted
-					latest_nightly = Some(line.clone());
-				}
-			}
-
-			latest_nightly?.trim_end_matches(&host).into()
-		},
-	};
-
-	Some(CargoCommand::new_with_args("rustup", &["run", &version, "cargo"]))
-}
-
-/// Wraps a specific command which represents a cargo invocation.
-#[derive(Debug)]
-struct CargoCommand {
-	program: String,
-	args: Vec<String>,
-}
-
-impl CargoCommand {
-	fn new(program: &str) -> Self {
-		CargoCommand { program: program.into(), args: Vec::new() }
-	}
-
-	fn new_with_args(program: &str, args: &[&str]) -> Self {
-		CargoCommand {
-			program: program.into(),
-			args: args.iter().map(ToString::to_string).collect(),
-		}
-	}
-
-	fn command(&self) -> Command {
-		let mut cmd = Command::new(&self.program);
-		cmd.args(&self.args);
-		cmd
-	}
-
-	/// Check if the supplied cargo command is a nightly version
-	fn is_nightly(&self) -> bool {
-		// `RUSTC_BOOTSTRAP` tells a stable compiler to behave like a nightly. So, when this env
-		// variable is set, we can assume that whatever rust compiler we have, it is a nightly
-		// compiler. For "more" information, see:
-		// https://github.com/rust-lang/rust/blob/fa0f7d0080d8e7e9eb20aa9cbf8013f96c81287f/src/libsyntax/feature_gate/check.rs#L891
-		env::var("RUSTC_BOOTSTRAP").is_ok() ||
-			self.command()
-				.arg("--version")
-				.output()
-				.map_err(|_| ())
-				.and_then(|o| String::from_utf8(o.stdout).map_err(|_| ()))
-				.unwrap_or_default()
-				.contains("-nightly")
-	}
-}
-
-/// Wraps a [`CargoCommand`] and the version of `rustc` the cargo command uses.
-struct CargoCommandVersioned {
-	command: CargoCommand,
-	version: String,
-}
-
-impl CargoCommandVersioned {
-	fn new(command: CargoCommand, version: String) -> Self {
-		Self { command, version }
-	}
-
-	/// Returns the `rustc` version.
-	fn rustc_version(&self) -> &str {
-		&self.version
-	}
-}
-
-impl std::ops::Deref for CargoCommandVersioned {
-	type Target = CargoCommand;
-
-	fn deref(&self) -> &CargoCommand {
-		&self.command
-	}
-}
-
-/// Returns `true` when color output is enabled.
-fn color_output_enabled() -> bool {
-	env::var(crate::WASM_BUILD_NO_COLOR).is_err()
 }
